Separate logging into separate file
